Divided server package from main
