+++ main.go
@@ -1,0 +1,15 @@
+package main
+
+import (
+	"github.com/ubiquitous-signage/hamster/server"	
+	// "github.com/ubiquitous-signage/hamster/lectures"
+	// "github.com/ubiquitous-signage/hamster/newsletters"
+	// "github.com/ubiquitous-signage/hamster/portal"
+	// "github.com/ubiquitous-signage/hamster/schedules"
+	// "github.com/ubiquitous-signage/hamster/trains"
+	// "github.com/ubiquitous-signage/hamster/weather"
+)
+
+func main() {
+	server.Run()
+}
