+++ panel/panel.go
@@ -1,0 +1,14 @@
+package panel
+
+type Panel struct {
+	Version  float64     `json:"version"`
+	Type     string      `json:"type"`
+	Title    string      `json:"title"`
+	Category string      `json:"category"`
+	Contents interface{} `json:"contents"`
+}
+
+type Content struct {
+	Type    string `json:"type"`
+	Payload string `json:"payload"`
+}
